+++ components/Section/Section.tsx
@@ -7,7 +7,7 @@ interface MyProps {
 export default function Section(props: MyProps) {
   return (
     <section>
-      <div className="mx-auto max-w-screen-xl px-10 pb-12 sm:px-10 lg:px-8">
+      <div className="mx-auto max-w-screen-xl px-6 pb-12 sm:px-10 lg:px-10">
         <div className="max-w-3xl">
           <SectionTitle
             title={"Metodologi"}
